Add BlueOak to acceptable licenses (#8472)
## Describe your changes

Our licenses check is failing because one of the frontend dependencies
changed the license to `BlueOak`. We already checked this with legal a
couple of weeks ago and its fine to add this to the acceptable licenses.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.