Add ability to clear cache for specific function arguments by passing args to `<cached_func>.clear()` (#8297)
## Describe your changes

Hi, I really love streamlit and have been having fun developing an LLM
app with it. I wanted to cache the LLM responses, unless a user said it
was a bad response. So I decided to implement this. Thanks for taking a
look!

I know it might not be perfect, if you give me some pointers I am happy
to make a lot of changes as needed :) I think the implementation with
passing the key though two layers of methods could be better so I'll
think how I can improve that.

This keeps the exisiting interface to CachedFunc.clear() where passing
no arguments clears all the cache, but also adds the option that if
args/kwargs are passed, the value key will be calculated and the cache
storage for that key will be invalidated. This only applies to DataCache
for now - I know the error raised could be better so please let me know
what type of exception you would prefer it to throw.


## GitHub Issue Link

[8153](https://github.com/streamlit/streamlit/issues/8153)

## Testing Plan

- ~~Explanation of why no additional tests are needed~~ **tests are
always needed**
- Unit Tests (JS and/or Python) **python unit test only**
- E2E Tests **not sure**
- Any manual testing needed? **since its very explicit I think unit
tests are better**

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.