Update address output in headless mode (#8647)
## Describe your changes

Closes #8629

When Streamlit is started via `server.headless=true`, the address output
does not contain `localhost`, although the server can still be reached
via `localhost`. The `headless` mode just defines whether a browser is
started, not what the address binding is. So, the idea is to show
`localhost` also in case when `server.headless=true` is set.


## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.