Don't allow writing widgets outside the fragment (#8756)
## Describe your changes

This PR adds a new check-policy function that must be called by every
widget (element that _can_ trigger a re-run).
Fragment-decorated functions - which includes dialogs - are not allowed
to write widgets to delta paths outside of the fragment's path itself.
We want to forbid this as such elements themselves can trigger re-runs
and we have seen weird issues such as widgets being duplicated with it.

For this, we store the `delta_path` of the current fragment when a
fragment-decorated code is executed. Otherwise, we wouldn't be able to
traverse an element's parent path and differentiate between a fragment
and non-fragment container.

The new check ensures that the delta path of an element starts with the
complete delta path of the fragment, if the element is inside of a
fragment.

## GitHub Issue Link (if applicable)

## Testing Plan

- Unit Tests (JS and/or Python)
- New unit tests are added. For this, different "apps" are combined with
all widgets and non-widgets and then asserted for success and failure.
In total, ~260 unit tests are executed with different combinations and
elements (~1.5 second in total)
- Any manual testing needed?
  - Manually tested this new check with a variety of apps

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.