Enforce that only the outermost command is tracked (#8769)
## Describe your changes

We have some functionality that ensures that only the outermost command
is tracked in metrics. Commands should not be tracked if they are part
of another higher-level command. However, this functionality was a bit
broken and only prevented the first inner command from being tracked.
This PR adds an additional flag to make sure that only the outermost
command can actually reset the global flag that prevents tracking from
inner commands.

## Testing Plan

- Improved unit test.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Lukas Masuch <103002573+sfc-gh-lmasuch@users.noreply.github.com>