Print bottom container always at the end of the app content (#9129)
## Describe your changes

In print-mode, the bottom container can overlay the actual app. For
example, a floating chat input might hide the app content. For this
reason, we want to move the bottom container to the very end when
printing.

## Testing Plan

- We have some printing e2e tests, but this one here relies on
multi-page content etc. This seems to be niche-enough to not have an
extra printing test just for this. I have left a comment at the
respective place to let future fellow devs know what the CSS-line was
added for.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.