Fix copy-buttons in dialogs (#9130)
## Describe your changes

Closes https://github.com/streamlit/streamlit/issues/9112

Right now, copy buttons in dialogs don't work. This seems to be because
the Modal swallows the focus events. This PR fixes this by using the
`container` property from `ClipboardJS` (inspiration from
[here](https://github.com/mac-s-g/react-json-view/issues/131#issuecomment-444095616)).
This is now also used in our `ReactJSON` element.

In the first iteration we set `focusLock={false}` on the Dialog which
also solved this issue. However, that came with an accessibility
drawback where you would be able to select elements outside of the
container via the `Tab` key.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
  - Add an e2e test to ensure that copying works
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.