Do not preserve children of a block if the block type changes (#9276)
## Describe your changes

We have this convenience that preserves the children of a block even if
a block changes. This was likely because we want to keep the same block
and children and we don't want to make random elements disappear (and
preserve widget state). This change keeps the existing behavior, but it
_will_ reset the children if the _type_ of the block changes as well.

## GitHub Issue Link (if applicable)
Closes #9259
Closes #8676

## Testing Plan

- Added a unit test to define the behavior accurately

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.