Unstale just updated fragments (#9285)
## Describe your changes

During a full app rerun, elements that are updated (have the same
scriptRunId as the current scriptRunId) are immediately shown as
un-stale. For fragment runs, the elements are shown as stale until the
end and then all at once are un-staled.
In this PR, we show elements as non-stale during the script run also for
fragments.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
  - update JS unit test to cover the updated behavior
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.