Remove mypy_custom_script.py (#9311)
## Describe your changes

In PR #9245, @lukasmasuch noticed that the mypy script (that invokes the
mypy program) was likely unnecessary, and we could probably just invoke
mypy directly. At the time, I assumed this would break something that
would then have to be fixed. But it seems that the replacement was
completely painless!

## Testing Plan

Checking the output of the CI on this, to make sure it's checking the
same number of files as the last commit (minus one, as we've just
removed one python file!) and reporting a success is a good step to make
sure the transition has been successful.

## Elaborate Digression

### The dream of what can be

So, what was the mypy script supposed to do, anyway, and why didn't it
do anything? Well, it seems to parse and then print some diagnostic
someone once thought was important. Internally, this passes
--lineprecision-report to mypy, which makes mypy make a text file
lineprecision.txt in a directory with a table like this:

```
Name                                                               Lines  Precise  Imprecise  Any  Empty  Unanalyzed
--------------------------------------------------------------------------------------------------------------------
02_App_Page_2                                                         17        2          0    0     15           0
03_App_Page_3                                                         17        2          0    0     15           0
audit_frontend_licenses                                              198      103          1    5     89           0
check_license_headers                                                105       21          0   13     71           0
cli_regression_tests                                                 216       33          1   73    109           0
```

And, if mypy_custom_script.py were working properly, this script would
process that text file and print out a table that looks like
[this](https://github.com/wyattscarpenter/streamlit/actions/runs/10470465177/job/28995682225#step:7:17):
```
Module                                                    Lines  Typed  Percent
------                                                    -----  -----  -------
02_App_Page_2                                                 2      2  100.00%
03_App_Page_3                                                 2      2  100.00%
audit_frontend_licenses                                     109    103   94.50%
check_license_headers                                        34     21   61.76%
cli_regression_tests                                        107     42   39.25%
```

It seems like `Lines` in this second table is `Lines` in the first,
minus `Empty`, and `Typed` in this table is `Precise` from the first
table.

### The burden of what has been

So why wasn't that report-printing code ever running? Well, it's
basically because mypy, which we call the main function of, calls
sys.exit (or
[something](https://github.com/python/mypy/blob/master/mypy/main.py#L60)
even worse) after it finishes, which kills our program as well, before
process_report can run. To change this, you have to
[change](https://github.com/wyattscarpenter/streamlit/commit/3acf200f1b2f98a45864f5240c10500e3ef4d25b)
```
mypy_main.main(stdout=sys.stdout, stderr=sys.stderr, args=args)
```
to
```
try:
    mypy_main.main(stdout=sys.stdout, stderr=sys.stderr, args=args, clean_exit=True)
except BaseException as e:
    pass
```
(It occurs to me that this might make the CI step pass in all
circumstances, because the exception is caught instead of crashing the
program, so the error code might be different. So, maybe the code would
have to be further altered in order to account for this.)

### Why delete the code if you can fix it?

It seems to me this project has been getting along pretty well without
this table! So, empirically, seems like no need for it anymore. (Maybe
someone will reply to this PR saying they would still like it!)

The way I like to do git histories, I would probably make a git commit
fixing this feature and then a git commit removing it, in case I ever
wanted to go back easily to a commit when it worked. But, on second
thought, this long description appearing in the git commit message for
the removal, discussing how one would fix the situation, is probably
just as good for that purpose.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.