Reformat some comments so that `make help` works. (#9317)
## Describe your changes

Previously, `make help`, which is supposed to give output like

```all                            Get dependencies, build frontend, install Streamlit into Python environment.```

 would emit some lines like

```build-deps the development process). Does not build the frontend.```.

Other lines were not formatted in a consistent way, and some commands
were not documented at all. This commit changes the source comments to
fix all that.

## Testing Plan

Although it's theoretically possible to conceive of a unit test to
prevent this situation, such an expenditure of effort strikes me as
unwise.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.